init data file clean

+++ code/v0/Source/executing/init_orientation.c
@@ -6,7 +6,7 @@
 /*   By: lazanett <lazanett@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/22 13:11:18 by lazanett          #+#    #+#             */
-/*   Updated: 2024/03/22 17:56:39 by lazanett         ###   ########.fr       */
+/*   Updated: 2024/03/25 14:31:05 by lazanett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,14 +18,14 @@ static void	init_orient2(t_all_data *all)
 	{
 		all->file->raydir.x = 0;
 		all->file->raydir.y = 1;
-		all->file->rayplane.x = 0.66;
+		all->file->rayplane.x = (long)0.66;
 		all->file->rayplane.y = 0;
 	}
 	if (all->world->start_angle == 78)
 	{
 		all->file->raydir.x = 0;
 		all->file->raydir.y = -1;
-		all->file->rayplane.x = -0.66;
+		all->file->rayplane.x = (long)-0.66;
 		all->file->rayplane.y = 0;
 	}
 }
@@ -39,14 +39,14 @@ void	init_orient(t_all_data *all)
 		all->file->raydir.x = -1;
 		all->file->raydir.y = 0;
 		all->file->rayplane.x = 0;
-		all->file->rayplane.y = 0.66;
+		all->file->rayplane.y = (long)0.66;
 	}
 	if (all->world->start_angle == 69)
 	{
 		all->file->raydir.x = 1;
 		all->file->raydir.y = 0;
 		all->file->rayplane.x = 0;
-		all->file->rayplane.y = -0.66;
+		all->file->rayplane.y = (long)-0.66;
 	}
 	init_orient2(all);
 }
